+++ module/party.js
@@ -1,39 +1,41 @@
 import { WwnPartySheet } from "./dialog/party-sheet.js";
 
+/**
+ * @param {import("../../foundryvtt/client/applications/sidebar/tabs/actor-directory.mjs").default} object
+ * @param {HTMLElement} html
+ */
 export const addControl = (object, html) => {
-  // Convert html to jQuery if it's not already
-  const $html = html instanceof jQuery ? html : $(html);
-
-  let control = `<button class='wwn-party-sheet' type="button" title='${game.i18n.localize('WWN.dialog.partysheet')}'><i class='fas fa-users'></i></button>`;
+  const control = document.createElement("button");
+  control.type = "button";
+  control.className = "wwn-party-sheet";
+  control.title = game.i18n.localize("WWN.dialog.partysheet");
+  control.innerHTML = `<i class="fas fa-users"></i>`;
+  control.addEventListener("click", (ev) => showPartySheet(object, ev));
 
   // Try to find the toggle-search-mode element first since we know it exists
-  const toggleSearch = $html.find(".toggle-search-mode");
-  if (toggleSearch.length) {
-    toggleSearch.before($(control));
+  const toggleSearch = html.querySelector(".toggle-search-mode");
+  if (toggleSearch) {
+    toggleSearch.before(control);
   } else {
     // Fallback to inserting at the start of the header
-    $html.find("#actors").find("header").prepend($(control));
+    html.querySelector("header")?.prepend(control);
   }
+};
 
-  $html.find('.wwn-party-sheet').click(ev => {
-    showPartySheet(object);
-  });
-}
-
-export const showPartySheet = (object) => {
-  event.preventDefault();
+export const showPartySheet = (object, event) => {
+  event?.preventDefault();
   new WwnPartySheet(object, {
     top: window.screen.height / 2 - 180,
     left: window.screen.width / 2 - 140,
   }).render(true);
-}
+};
 
-export const update = (actor, data) => {
-  if (actor.getFlag('wwn', 'party')) {
-    Object.values(ui.windows).forEach(w => {
+export const update = (actor) => {
+  if (actor.getFlag("wwn", "party")) {
+    Object.values(ui.windows).forEach((w) => {
       if (w instanceof WwnPartySheet) {
         w.render(true);
       }
-    })
+    });
   }
-}
+};